S02P13A208-38 S02P13A208-36 S02P13A208-47 filtering&input_date_timezone&etc modify

+++ backend/heartmarket/src/main/java/com/heartmarket/model/service/MannerServiceImpl.java
@@ -107,15 +107,15 @@ public class MannerServiceImpl implements MannerService{
 	}
 	
 	@Override
-	public ResultMap<Map<String, Double>> findManner(String email){
+	public ResultMap<Map<String, Integer>> findManner(String email){
 		try {
 			Manner manner = mr.findBymUserUserNo(ur.findByEmail(email).getUserNo());
 			if(manner != null) {
-				Map<String, Double> rm = new HashMap<String, Double>();
-				rm.put("heartgauge", manner.getHeartGauge());
-				return new ResultMap<Map<String, Double>>("SUCCESS", "매너 불러오기", rm);
+				Map<String, Integer> rm = new HashMap<String, Integer>();
+				rm.put("heartgauge",(int) manner.getHeartGauge());
+				return new ResultMap<Map<String, Integer>>("SUCCESS", "매너 불러오기", rm);
 			}else {
-				return new ResultMap<Map<String, Double>>("FAIL", "찾을 수 없는 유저입니다.", null);
+				return new ResultMap<Map<String, Integer>>("FAIL", "찾을 수 없는 유저입니다.", null);
 			}
 		}catch(Exception e) {
 			e.printStackTrace();
